# openapi.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Work Manager
  description: >-
    Customizable work item manager that leverages Camunda workflows (BPMN and DMN), Form.io and customizable data types
    to model online transaction processing.
paths:
...
  /transaction:
    post:
      summary: Create a new transaction instance.
      operationId: postTransaction
      tags: [ 'Transaction' ]
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        description: Transaction Creation Request.

        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreationRequest'
      responses:
        200:
          description: The posted transaction  post-save. This will include assigned IDs for newly created transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionModel'
    get:
      summary: Get list of transactions for a given transaction definition.
      operationId: getTransactions
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: transactionDefinitionKey
          description: Key of the transaction definition to fetch transactions for.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Collection of transaction matching definition.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionModel'
  /transaction/search:
    get:
      summary: Get filtered list of transactions
      operationId: getFilteredTransactions
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: transactionDefinitionKey
          description: Key of the transaction definition to fetch transactions for.
          required: false
          schema:
            type: string
        - in: query
          name: category
          description: category key to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: startDate
          description: start date to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: endDate
          description: end date to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: priority
          description: priority to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: status
          description: status to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: assignedTo
          description: assigned user to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: assignedToMe
          description: filter transactions assigned only to yourself (will ignore the value of assignedTo if true).
          required: false
          schema:
            type: boolean
        - in: query
          name: sortCol
          description: the column to sort the transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: sortDir
          description: the direction to sort the transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: pageNumber
          description: the page number to show transactions for (default = 0).
          required: false
          schema:
            type: integer
        - in: query
          name: pageSize
          description: the number of transactions to retrieve (default = 25).
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Collection of transaction matching filter criteria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTransactionModel'
  /transaction/category:
    get:
      summary: Get list of transactions by transaction category.
      operationId: getTransactionsByCategory
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: category
          description: category key to filter transactions by.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Collection of transaction matching definition.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionModel'
  /transaction/{id}:
    get:
      summary: Get a single transaction by ID.
      operationId: getTransaction
      tags: [ 'Transaction' ]
      parameters:
        - in: path
          name: id
          description: ID of transaction to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionModel'
        '404':
          description: A transaction with the given ID does not exist.
    post:
      summary: Update an existing transaction.
      operationId: updateTransaction
      tags: [ 'Transaction' ]
      parameters:
        - in: path
          name: id
          description: ID of transaction to update
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: taskId
          description: (Optional) ID of workflow task to post data to as part of this update
          required: false
          schema:
            type: string
      requestBody:
        description: Transaction Data to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateRequest'
      responses:
        '200':
          description: The updated transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionModel'
...
  /transaction/link:
    post:
      summary: Create a link between two transactions
      operationId: linkTransactions
      tags: [ 'Transaction' ]
      requestBody:
        description: The transaction link creation request.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLinkCreationRequest'
      responses:
        201:
          description: The created transaction link id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLinkModel'
  /transaction/linkedtransactions/{id}:
    get:
      summary: Get the linked transactions for a given transaction id
      operationId: getLinkedTransactionsById
      tags: [ 'Transaction' ]
      parameters:
        - in: path
          name: id
          description: The transaction id to search on.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: The list of linked transactions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkedTransaction'
  /transaction/statuses:
    get:
      summary: Get list of available statuses defined in Camunda workflow
      operationId: getAvailableStatuses
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: type
          description: The type of statuses to retrieve (public or internal).  Default is public.
          schema:
            type: string
            enum: [ public, internal ]
        - in: query
          name: category
          description: (Optional) The transaction definition category to search for
          schema:
            type: string
        - in: query
          name: key
          description: (Optional) The transaction definition key to search for
          schema:
            type: string
      responses:
        200:
          description: The list of available statuses.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /transaction/statuses/count:
    get:
      summary: Get list of statuses with a count of how many transactions have each status
      operationId: getTransactionCountByStatus
      tags: ['Transaction' ]
      parameters:
        - in: query
          name: transactionDefinitionKey
          description: Key of the transaction definition to fetch transactions for.
          required: false
          schema:
            type: string
        - in: query
          name: category
          description: category key to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: startDate
          description: start date to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: endDate
          description: end date to filter transactions by.
          required: false
          schema:
            type: string
        - in: query
          name: priority
          description: priority to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: status
          description: status to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: assignedTo
          description: assigned user to filter transactions by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: assignedToMe
          description: filter transactions assigned only to yourself (will ignore the value of assignedTo if true).
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Collection of transaction counts by status matching filter criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionCountByStatusModel'
...
components:
  schemas:
...
    TransactionCreationRequest:
      type: object
      required:
        - transactionDefinitionKey
      properties:
        transactionDefinitionKey:
          type: string
    TransactionUpdateRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          additionalProperties: true
        condition:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
    TransactionCountByStatusModel:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
    PagedTransactionModel:
      type: object
      properties:
        totalPages:
          type: integer
        totalCount:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionModel'
    TransactionModel:
      type: object
      required:
        - data
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        priority:
          type: string
        district:
          type: string
        transactionDefinitionId:
          type: string
          format: uuid
        transactionDefinitionKey:
          type: string
        processInstanceId:
          type: string
        createdBy:
          type: string
        assignedTo:
          type: string
        createdByDisplayName:
          type: string
        subjectUserId:
          type: string
        subjectUserDisplayName:
          type: string
        createdTimestamp:
          type: string
          format: date-time
        lastUpdatedTimestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
    TransactionLinkTypeModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        fromDescription:
          type: string
        toDescription:
          type: string
...
    TransactionLinkModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromTransactionId:
          type: string
          format: uuid
        toTransactionId:
          type: string
          format: uuid
        transactionLinkType:
          $ref: '#/components/schemas/TransactionLinkTypeModel'
    TransactionLinkCreationRequest:
      type: object
      properties:
        fromTransactionId:
          type: string
          format: uuid
        toTransactionId:
          type: string
          format: uuid
        transactionLinkTypeId:
          type: string
          format: uuid
    LinkedTransaction:
      type: object
      properties:
        description:
          type: string
        linkedTransactionId:
          type: string
          format: uuid
...